requirements.txt
ccxt
sanic
pyyaml
aiohttp
pandas
cerberus
pytest
pytest-asyncio
pytest-sanic
